fix warnings

+++ rd/read-tiff.c
@@ -55,7 +55,7 @@ tiff_init(FILE *fp, char *filename, unsigned int page,
 		"nsamples=%d, depth=%d fo=%d pm=%d scanline=%" PRId32 "\n",
 		h->width,h->height,h->config,h->nsamples,h->depth,
 		h->fillorder,h->photometric,
-		TIFFScanlineSize(h->tif));
+		(uint32_t)TIFFScanlineSize(h->tif));
 
     if (PHOTOMETRIC_PALETTE   == h->photometric  ||
 	PHOTOMETRIC_YCBCR     == h->photometric  ||
